+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/recovery/RecoveryLivenessTest.java
@@ -17,7 +17,7 @@
 
 package com.radixdlt.integration.recovery;
 
-import com.radixdlt.client.ValidatorAddress;
+import com.radixdlt.identifiers.ValidatorAddress;
 import com.radixdlt.ledger.LedgerAccumulator;
 import com.radixdlt.ledger.SimpleLedgerAccumulatorAndVerifier;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
