+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/JsonRpcUtil.java
@@ -30,7 +30,7 @@ import java.util.function.Function;
 import java.util.function.Predicate;
 
 import static com.radixdlt.client.api.ApiErrors.MISSING_PARAMS;
-import static com.radixdlt.utils.CommonErrors.UNABLE_TO_DECODE;
+import static com.radixdlt.identifiers.CommonErrors.UNABLE_TO_DECODE;
 import static com.radixdlt.utils.functional.Failure.failure;
 import static com.radixdlt.utils.functional.Result.fromOptional;
