+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/system/SystemController.java
@@ -19,7 +19,7 @@
 package com.radixdlt.api.system;
 
 import com.radixdlt.DefaultSerialization;
-import com.radixdlt.client.ValidatorAddress;
+import com.radixdlt.identifiers.ValidatorAddress;
 import com.radixdlt.ledger.VerifiedTxnsAndProof;
 import com.radixdlt.network.addressbook.AddressBook;
 import com.radixdlt.network.addressbook.PeerWithSystem;
