+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/client/api/ApiErrors.java
@@ -20,16 +20,11 @@ package com.radixdlt.client.api;
 import com.radixdlt.utils.functional.Failure;
 
 public enum ApiErrors implements Failure {
-	BASE_URL_IS_MANDATORY(2501, "Base URL is mandatory"),
-	NOT_A_JSON_OBJECT(2502, "Not an JSON object"),
-	NO_CONTENT(2503, "No content in response"),
 	SYMBOL_DOES_NOT_MATCH(2504, "Symbol {0} does not match"),
 	INVALID_PAGE_SIZE(2505, "Size {0} must be greater than zero"),
 	MISSING_PARAMS(2506, "The 'params' field must be present and must be a JSON array"),
 	INVALID_NETWORK_ID(2507, "Network ID is not an integer"),
 	UNKNOWN_VALIDATOR(2508, "Validator {0} not found"),
-	INVALID_VALIDATOR_ADDRESS(2509, "Invalid validator address {0}"),
-	INVALID_ACCOUNT_ADDRESS(2510, "Invalid account address {0}"),
 	INVALID_BLOB(2511, "Invalid blob {0}"),
 	INVALID_SIGNATURE_DER(2512, "Invalid signature DER {0}"),
 	INVALID_PUBLIC_KEY(2513, "Invalid public key {0}"),
