+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/client/handler/ActionParserTest.java
@@ -17,8 +17,8 @@
 
 package com.radixdlt.client.handler;
 
-import com.radixdlt.client.AccountAddress;
-import com.radixdlt.client.ValidatorAddress;
+import com.radixdlt.identifiers.AccountAddress;
+import com.radixdlt.identifiers.ValidatorAddress;
 import com.radixdlt.client.store.ClientApiStore;
 import com.radixdlt.utils.functional.Result;
