+++ radixdlt-java-common/src/main/java/com/radixdlt/utils/UInt256.java
@@ -26,7 +26,7 @@ import java.math.BigInteger;
 import java.util.Arrays;
 import java.util.Objects;
 
-import static com.radixdlt.utils.CommonErrors.UNABLE_TO_DECODE;
+import static com.radixdlt.identifiers.CommonErrors.UNABLE_TO_DECODE;
 
 /**
  * A 256-bit unsigned integer, with comparison and some basic arithmetic
