+++ radixdlt-java-common/src/main/java/com/radixdlt/utils/UInt384.java
@@ -24,7 +24,7 @@ import com.radixdlt.utils.functional.Result;
 import java.util.Arrays;
 import java.util.Objects;
 
-import static com.radixdlt.utils.CommonErrors.UNABLE_TO_DECODE;
+import static com.radixdlt.identifiers.CommonErrors.UNABLE_TO_DECODE;
 
 /**
  * A 384-bit unsigned integer, with comparison and some basic arithmetic
