+++ radixdlt-java/radixdlt-java/src/main/java/com/radixdlt/client/lib/ClientLibErrors.java
@@ -15,22 +15,23 @@
  * language governing permissions and limitations under the License.
  */
 
-package com.radixdlt.utils;
+package com.radixdlt.client.lib;
 
 import com.radixdlt.utils.functional.Failure;
 
-public enum CommonErrors implements Failure {
-	AID_IS_NULL(1601, "AID string is 'null'"),
-	INVALID_LENGTH(1602, "AID string has incorrect length {0}"),
-	UNABLE_TO_DECODE(1603, "Unable to decode: {0}"),
-	UNABLE_TO_DESERIALIZE(1604, "Unable to deserialize: {0}"),
-	CANT_MAKE_RECOVERABLE(1701, "Unable to convert signature to recoverable {0}"),
-	INVALID_RADIX_ADDRESS(1702, "Invalid RadixAddress {0}");
+public enum ClientLibErrors implements Failure {
+	BASE_URL_IS_MANDATORY(1001, "Base URL is mandatory"),
+	NOT_A_JSON_OBJECT(1002, "Not an JSON object"),
+	NO_CONTENT(1003, "No content in response"),
+	INVALID_NETWORK_ID(1004, "Network ID {0} is not an integer"),
+	MISSING_NETWORK_ID(1005, "Network ID not found"),
+	MISSING_FIELD(1006, "Field {0} is missing"),
+	CONNECTION_ERROR(1007, "Error: {0}");
 
 	private final int code;
 	private final String message;
 
-	CommonErrors(int code, String message) {
+	ClientLibErrors(int code, String message) {
 		this.code = code;
 		this.message = message;
 	}
